Add ObjNum and the ability to call methods on numbers
Since numbers are special (being encoded as, well, numbers as opposed to
pointers like literally everything else) the system for calling methods
on them and defining C++ method implementations is quite specialised.

This is a big pile of magic, and ObjNumClass has special handling inside
the bindings generator script.

The basic idea is that we define the class that all numbers are a type
of (ObjNumClass) without an ObjNum that actually instanciates it. Any
instance methods on the ObjNumClass class will become instance methods
on individual numbers, while static methods on ObjNumClass will become
static Num methods (eg Num.maxSafeInteger). This isn't how any other
classes work, and hence the modifications to the bindings generator
script.

P.S: It's probably a very good idea to prevent classes from extending
Num since the results could be *very* weird, and almost certainly hard
to debug for most users.