Remove ExprLogicalNot
This was only used to invert jump conditions, and now we have the
jumpIfFalse flag we can just use that.

This node was ambiguous as to whether it'd evaluate the expression and
invert it, or call the '!(_)' method on it's target. Calling the method
isn't something we'd want to do as part of an implementation detail and
would be slow, so it's easier to just get rid of this node.