Add field support
This is done by allocating a bit of extra memory at the each of every
ObjManaged that has fields. The fields for superclasses are stored
before subclasses, similar to C++ or OpenJDK.

A problem is that we don't know ahead of time how large an object will
be. This is because a class can extend (well, extending is implemented
yet, but this is how it will be once it is) a class from another file,
compiled separately. Thus we don't know at compile time where our fields
are stored relative to the start of the object.

The solution is to have global variables for each class, which store
this offset. During module initialisation these variables are
initialised by calling a method in the runtime. This also conveniently
lets us modify the layout of ObjManaged without an ABI break.

In the future when doing more optimisations we might try and figure out
the size of superclasses if possible to avoid having to go through this
global, but for now it should be more than good enough.

At runtime, at the start of each method this variable is loaded, on the
presumption that a field is used at some point in the method. This isn't
always the case, and we could optimise things by checking if a field is
actually used before loading it.