Convert upvalue imports from statements to pure variables
Previously upvalues took the form of StmtUpvalueImport, which was both a
variable and a statement. The idea was this statement would require that
a variable was closed over if the parent function returns.

This probably isn't going to do us any good anyway (what happens if the
child function only runs that statement after the parent returns?) and
thus I've just made it a normal variable.

There may well be some merit in having a statement to declare that it
won't use upvalues anymore before the function returns, but that's not
what this was, and we can put it in later.