Initial and incomplete upvalue closing
This is the initial and very incomplete implementation of upvalue
closing/relocation, where variables referenced by closures are moved to
the heap once the scope that owns them has returned.

There are two huge problems with the current implementation - it
basically doesn't work in loops (all closures will see the variable in
the last iteration, as we don't track which functions we've already
fixed up) and two closures that use the same variables will each see
their own copy of it.