Fix returns after dead labels not getting removed
If you had something like this:

myMethod() {
  if (a) {
    return 1;
  } else {
    return 2;
  }
}

It would fail to compile, as there would be two back-to-back return
statements at the end of the function - return(2) and return(null).

This is because there would be a label to jump to the end of the
function (and thus to the implicit return(null) statement) that was used
by a jump after return(1) statement. This jump would be removed, causing
the label to be 'dead' and get deleted. However this happens after the
point where anything-after-a-return is removed, so the return(null) was
left in.

This patch just runs the cleanup code for the single block again after
removing labels from it.