Make the compiler not depend on the runtime library
This moves shared stuff line NaN-packing and function hashing into a new
static library that's linked into both the compiler and runtime library.

In the future this will let us use the Wren compiler to compile parts of
the runtime library, particularly to reuse Wren's wren_core.wren script
which defines a bunch of functions on the core classes.