Make rtlib's symbols hidden by default
This is in preparation for linking wren_core.wren into the runtime
library. We need to hide symbols otherwise we'd have to add support for
the GOT in our compiler, which is a pain and quite unnecessary at this
point (with the exception of the true and false constants, but we'll
handle them separately).