Stop corrupting the stack with locals
All variables that can be used in upvalues are stored in a special
location on the stack. To access them, take the variable ID and multiply
it by the size of a Value (a 64-bit value, the size that had been
allocated for them).

Due to an unfortunate typo (VLocal instead of Value), these offsets were
previously multiples of 32 bytes instead, corrupting the region just
above the current stackframe, which was annoying to debug.