llvm-backend: Mark wren_get_core_class_value as pure
This indicates to the optimiser that this function can be removed or
called at any point without changing the programme's semantics. This is
mostly useful for dead-code elimination.

For this single function it's not useful (we only include variables that
are used), but we'll use the same attributes for other stuff too.