llvm-backend: Always heap-allocate upvalue storage
Previously, we'd handle upvalues similarly to upstream Wren, and the QBE
backend: all locals went on the heap, and if they were used by a closure
when a function ends, then they were relocated to the heap.

Performing this relocation was already on the borderline of intolerably
complicated for regular upvalues, but with nested closures and upvalues
that go through several closures, there wasn't any obvious solution to
do this in a performant manner.

Thus locals that are used as upvalues are always allocated on the heap,
and their address never changes which removes a lot of brittle and
complicated logic.