Get access in C++ to the wren_core globals
This adds an API to the QBE-generated code to retrieve the globals that
module defines. We'll use this in the future (also adding LLVMBackend
support) to implement imports, but for now we want it to import Wren
classes in the C++ runtime.

This is needed specifically for the upcoming attribute support - the
'attattributes' property of classes is itself a class defined fully in
Wren, and (since the Wren code uses fields) it'd be a big pain to try
and move it to C++. Thus we'll just instantiate the Wren class.