llvm-backend: Limit how string literals are used as IR names
Previously any string literal would be copied into the symbol names of a
couple of global variables. This is pretty obviously a bad idea - very
long strings would be fun - but it kept working until a specific
test [1] tried using a null byte in it's string literal, which tripped
an assertion in LLVM.

Now we limit these strings to 30 bytes, and remove any nulls.

[1] test/language/string/byte_escapes.wren