Allow specifying multiple module names on the command line
Previously, you could only specify a single module name, which was used
to compile all of the files provided. Since each file needs it's own
module name, this patch allows you to specify multiple module names.

It's a bit of an ugly system at the moment - the module names are
correlated with the order the source file names are specified - but
it'll let us get the basic module stuff going.