llvm-backend: Add initial GC support
This uses LLVM's RS4GC pass to spill all the registers onto the stack
before (almost, there's a few exceptions) every function call. This
means we can unwind the stack and see/modify all the values for every
Wren function in the stack trace.

The offsets of the stack (relative to RSP - this involves a very small
amount of x86-specific code) to access each value for a bunch of given
instruction pointers is written to a new block, a pointer to which is
stored in the global variables array.

This is enough compile-time support to implement a copying GC in the
runtime - that will very much come as it's own commit later on.