Run the main module by exporting it's globals function
Previously we exported a pointer to the main function, which the
executable stub picked up on to run the user's code. We now export a
pointer to the module getter function.

This will be useful for GC work, when we need to use the stackmaps from
all the available modules.