test: Add expected failures support
There's quite a few tests left that still fail because we don't yet
support them. However, the bulk of the test suite that passes is still a
useful check against regressions.

To differentiate tests that we're expecting to see fail because they
always have been and we haven't implemented them yet, and those that now
fail because of a regression, this patch marks those tests currently
failing as expected to fail. These tests show up in yellow and don't
have the full error messages printed out, to make regression failures
more noticeable.