IRVisitor: Allow visitors to substitute nodes
This patch passes references to pointers into the Visit function,
allowing the visitor to change what node a pointer is pointing to.

This allows visitors to substitute an expression or statement for
another, without having to have special logic for all the nodes that
might have pointers to those nodes.