IRCleanup: Flatten out calls and StmtEvalAndIgnore
Previously, statements could be nested inside expressions using
StmtEvalAndIgnore. For analysis and optimisation this isn't very nice,
so we flatten it out to run before the statement it's used for.

This means we need to ensure that a call can't be nested in the
expressions of another call. Consider the following:

myFunc(a(), [b()])

Here, [b()] is a StmtEvalAndIgnore, so it'd get moved to a temporary
variable:

var tmp = [b()]
myFunc(a(), tmp)

And now b() is executed before a(). If these functions have side-effects
then the observable behaviour of the programme is changed. To solve
this, we move all function calls out to temporary variables whereever we
find them, in the order we find them:

var tmp1 = a()
var tmp2 = b()
var tmp3 = [tmp2]
myFunc(tmp1, tmp3)

Obviously this is going to create a lot of local variables, so we'll
need to handle them properly from a performance perspective. This is the
form that ultimately ends up in assembly anyway, we just have to make
sure these temporary variables don't clog up systems for handling user
variables (in particular, moving them to SSA form - they're only
assigned one, and we should make use of that).