llvm-backend: Wrap i64->value conversions in a function during RS4GC
RS4GC doesn't like it when we violate it's higher-level machine model
with inttoptr casts, as described in it's documentation. It was getting
confused about whether a value was derived from another, in a test as
simple as this:

for (i in 0..2) {
}

This temporaraly wraps all i64-to-'ptr addrspace(1)' casts in a function
that doesn't exist and has no implementation. This makes RS4GC happy,
and we remove the function afterwards, replacing it with the real value.