Add the per-object GC marking methods
This adds a new pure virtual method to Obj, MarkGCValues. This is called
by the GC to figure out what objects are reachable via a given object.

The current implementations are simple, but for the most part not
tested. We'll need to find some way to integrate GC testing into the
test suite.

This patch doesn't add the GC code that calls this method: that will be
added in a later patch, as this patch is already rather large.