Pass the module globals function when creating strings
This adds a new argument to wren_init_string_literal, namely the
get_globals function. This will later be used by the runtime to
associate string literals with modules to mark them as GC roots.

We could just put all string literals into a single big list and mark
them all as roots, but we'd never be able to implement module unloading
without a memory leak.

Since this is a big ABI break, doing it now is much better than waiting
until later.