Implement static fields for declared-once classes
This implements static fields backed by global variables. This works
great (and is certainly optimal from a performance perspective) for
classes which we know are only ever defined at most once.

This is the only type of class we support at the moment, but later on we
should add support for declaring classes in loops, functions and other
places like that, since upstream Wren does.

In those cases we'll need to store the static fields in the ObjClass
instances, since we'll presumably have one of those per declaration.