Add a node to indicate upvalues are coming into scope
Previously, the LLVM backend (forgetting the QBE one - that was even
more badly broken) allocated memory for upvalues at the start of a
function call. But that means variables declared in a loop aren't
independent across iterations - they're all treated as a single
variable.

This patch takes the first step to fix that: an IR node is generated in
the block where these variables come into scope, and later we'll modify
the backends to allocate memory at those locations.