qbe-backend: Throw away the upvalue relocation code
For now we just won't support upvalues on the QBE backend. The approach
used there was fundamentally a bad fit for this compiler, was broken for
all but trivial cases, and needed rewriting anyway.

With a little careful optimisation, allocating memory of a fixed size
from a slab allocator to store the upvalues is probably a lot faster
than our linked-list-per-closure monstrosity, and will be a lot more
robust and easier to maintain too.