llvm-backend: Store per-variable data with BackendNodeData
Previously we used a couple of maps to store this information, but using
the backend node data is both faster and cleaner.

In the future commit that uses multiple upvalue packs, this will keep
things a lot cleaner due to the increased number of per-variable
information we need to track.