Track the top-level StmtBeginUpvalues in each IRFn
This will be used by the upcoming changes to the LLVM backend, where it
generates multiple upvalue packs per function, corresponding to the
scopes locals are declared in.

It needs to know which upvalue pack is the root-level one, as that one
must be initialised prior to storing the function arguments into local
variables.