llvm-backend: Allocate separate upvalue packs per scope
Previously, the entire function had one upvalue pack for all variables.
However, this meant that multiple instances of a local variable would
all share a single slot and thus overwrite each other. For example, in
test/language/for/closure_in_body.wren:

var list = []

for (i in [1, 2, 3]) {
  var j = i + 1
  list.add(Fn.new { System.print(j) })
}

for (f in list) f.call()

This is supposed to print 2,3,4 but would previously print 4,4,4 as j
was being overwritten in each iteration of the loop, despite the fact it
should have become independent when the scope ends.

This patch means that in each scope where upvalues are declared, we'll
allocate a new storage block and store the variables belonging to that
scope there.