llvm-backend: Build objects as relocatable
This avoids using references to constants. This shouldn't have much
effect on the generated code, since practically all the times something
from a read-only section is used is during initialisation.

This change was motivated by the need to compile wren_core with this
option, but is useful any times you want to include a Wren module in a
shared library.