Properly handle ObjMap iterators in GC scanning
Our map's iterator value returns keys, and to prevent null or false from
ending the loop we set the LSB of their pointers.

This makes the GC scanner ignore that set LSB bit. This is a bit of a
hacky solution, but maps are probably integral enough to justify this.