Add the pre_entry_gc meson option, for GC testing
This adds an option to run the GC at the start of every C++ function
call. Using this while running the test suite will help uncover cases
where the GC is freeing reachable objects.

Obviously the performance overhead from this is considerable,
particularly with not-trivially-small heaps, and it's thus very much
only a testing/diagnostic option.