ObjFn: Mark the number of storage blocks in the upvalue pack
This adds something we'll need for freeing the storage blocks where the
upvalues live: the ability to allocate some extra memory in the upvalue
pack to hold the pointers to those storage blocks.

We'll use reference counting to free those storage blocks, so when
closures are destroyed they need to decrement the reference counts. The
slots in the upvalue pack are also used to pass the pointer to a storage
block from one closure to another (when an upvalue uses a variable
that's further out than the immediately enclosing function).

For now this doesn't actually do anything with those slots, it just
reserves them for the runtime. Neither backend requests any of them as
of this patch.