Support classes with lots and lots of functions
Normally, functions are stored in an inline 256-entry hash table.
However if a class has too many methods (eg the many_methods.wren test),
the table would run out of space.

The solution here is to use an STL unordered_map to store the entries
for classes with too many methods. The heuristic for switching to this
container is attempting to find an entry in the hashtable and finding
five filled slots before we find an empty one. This suggests the table
is overloaded and won't perform well during method lookups, at which
point moving to the STL map is probably warranted.