Implement fibre abort catching (with try)
When a fibre throws an error, it uses an exception to unwind the stack
and mark the fibre as failed. The fibre that called it will re-throw
that error, and that continues until either the root fibre fails and the
programme exists, or a fibre that called try() is hit.