Block fibres from being call()ed when they're waiting for another fibre
Previously, there was a single state for suspended fibres, and that was
used both for fibres that called yield() and for fibres that called
call() on another fibre and are waiting for them to yield.

The problem is that one musn't be able to call() a fibre that's in the
second state, since it's also supposed to be resumed when it's call()
finishes.

This patch thus adds a WAITING state, which is used for fibres that have
called call(), and disallows calling fibres that are in this state.