Implement a runner to run the API tests
This puts all the previously-committed pieces together, and lets you run
the API tests via the test runner.

There's a new executable that contains the C side of the test logic. The
Wren code gets built as a shared library and dynamically loaded by that
executable.

When using the test runner, this is done transparently.