Mark a tonne of stuff in the compiler as DLL_EXPORT
On Windows, to make llvm_backend a dynamic-link library (which is
particularly important on Windows, as LLVM can't build shared libraries
there under MSVC) we need to mark specific symbols as exported.

This marks a heap of stuff as exported, which should cover everything we
presently use, and hopefully also what we're likely to use in future
changes.