win: Link LLVM statically on Windows
You can't build LLVM as DLLs with MSVC, so this patch makes it link
statically on Windows.

It also turns the LLVM plugin (with the GC support) into a static
library, since it doesn't make sense to build it as a DLL (we'd have two
copies of large parts of LLVM, and AFAIK on Windows/MSVC LLVM doesn't
support loadable modules (see the --load flag in opt) anyway).