Add the initial skeleton of the Windows tinylink linker
The idea is that for Windows, we'll have a small custom linker that can
link Wren-only binaries.

We can't really do what we did on Linux with a partially-linked file
that we feed back into a proper linker, but the idea here is similar:
build the stub with all the C linking done by the MSVC linker, and only
link up the Wren modules in our custom linker.

This also means people won't have to install MSVC to get it's linker.