tinylink: Restructure everything for object merging
The idea is to allow us to combine multiple object's (for example) .text
sections into the same section in the output image.

There's probably still a lot of bugs - support for actually using it
with multiple objects isn't implemented - but we need this structure in
place to be able to make it work.