win: Fix exception handling on fibres
Windows keeps track of the start and end of it's stacks to detect stack
pivoting attacks, so we have to update these variables when we switch to
our own stacks.

Having the wrong values set can cause the C runtime to kill the process
when certain functions are called, and the exception handling code won't
dispatch any SEH exceptions (and thus no C++ exceptions) if the stack is
'wrong'.