tinylink: Fix .bss containing the PE header
OBJ files set a in-file size for the .bss section since that's the only
way they can indicate the size of a section. Tinylink wrote them out as
a normal section, and since the data-in-file pointer was set to zero, it
wrote out the object file into the .bss section.