Disable some very-slow-to-compile tests
These tests seem to interact very poorly with the RS4GC pass, causing
the compiler to get stuck (it was working, but didn't finish within a
couple of minutes on my desktop).

This is done before the subsequent test enabling RS4GC with the new
flag, so that you can git bisect over the codebase without getting
stuck.