When in the GC test mode, mark function arguments as roots
There's a mode that runs the GC every time a C++ function is called.
However, this didn't mark the function arguments as roots - which caused
an enormous number of spurious errors, as Wren code doesn't mark values
as GC roots when it's passed them into a function and doesn't use them
after that.

This is perfectly reasonable behaviour in most cases - when calling
other Wren functions or making calls to the C FFI the arguments are
correctly marked, and the System.gc() call doesn't take any arguments.
However, for this test mode, it becomes a huge problem.

When we start running the GC automatically, we'll need to find some way
to solve this - maybe by setting a flag that gets read on function
returns, to reduce the performance penalty of dealing with these
arguments 'properly'?