IRPrinter: Use properly-calculated colours for basic blocks
Previously, we used the hash of the block pointer to select a colour,
which often lead to multiple blocks with a very similar colour.

This assigns blocks different colours based on the order in which blocks
have their colour picked, which isn't ideal but is at least quick to
implement and much better than before.