llvm-backend: Support only basic-block inputs, remove fallthrough blocks
This avoids having the fallthrough blocks after conditional jumps, which
existed because previously the jump statement handler wasn't able to
read the following statement, and had to assume it might be something
other than a jump.

By only accepting input in basic block form, we can get it that
additional information to remove these useless blocks.

At the moment it has very little benefit, but in the future it might be
quite useful for Phi nodes.