Switch variable arguments to SSAVariable
This is in preparation for implementing a local-to-SSA pass, which will
be the basis for further optimisations (most notably, type analysis).

Having locals around can severely damage these optimisations, since
they'll all be operating on SSA variables. Thus we want to convert as
much as possible to using SSA variables.

The frontend immediately creates locals and assigns them to the SSA
variables. This makes them behave like all the other user-generated
locals, which the pass will have to handle.