Add initial type deduction support
This figures out what types variables are based on the results of
previous functions.

It's conservative but accurate - there's a lot of reasons why it might
not be able to figure out the type of a variable, but if it does it is
certainly that type.

We'll later use this to optimise function calls when we know the
underlying types, and in particular special-case basic number functions.