Add support for intrinsic number functions
This adds support for marking core class methods as intrinsics, and that
information will be made available to the backend.

This patch uses it to reduce number operations to fadd/fsub/fmul/fdiv
operations in the LLVM backend, though that implementation uses a LOT of
bitcasts that prevent optimisations like constant folding, which will be
the subject of a subsequent patch.