ide: Replace the iterative tree-walker with a recursive one
This significantly simplifies a lot of the node handling by moving the
node-processing logic to it's own functions. We could actually switch
back to an iterative implementation and it wouldn't be too bad, but it
seems somewhat unlikely that would be a major issue.

More importantly, each scope is now built separately, so we should be
able to re-build and substitute them in reliably, when we're
implementing edit updates.