ide: Build hashes for each scope, so we can later reuse them.
This hashes the contents (symbol-wise) of each scope, so we can identify
which scopes changed.

Each scope is now parsed separately, and doesn't parse any child scopes.
That is all done by a new loop in the update function, that parses all
the scopes. That can later be changed to re-use scopes where possible.